refactor: separate customized css config
