+++ src/styles/index.ts
@@ -1,5 +1,3 @@
 export * from './components';
-export { iconSizeVariants } from './icon';
 export * from './layout';
 export * from './pages';
-export { rootClassName } from './root';
