--- src/styles/root.ts
@@ -1,9 +1,0 @@
-import { css } from 'styled-system/css';
-
-export const rootClassName = css({
-	fontFamily: 'pretendard',
-	bg: 'background',
-	color: 'text',
-	transition: 'background 0.2s ease-in',
-	scrollBehavior: 'smooth',
-});
